Moved download functionality to actions
