+++ app/js/actions/members.js
@@ -10,6 +10,32 @@ function getMembersSuccess(members) {
   };
 }
 
+function saveAs(contents, filename) {
+  const element = document.createElement('a');
+  element.setAttribute('href', 'data:text/plain;charset=utf-8,' + encodeURIComponent(contents));
+  element.setAttribute('download', filename);
+  element.style.display = 'none';
+  document.body.appendChild(element);
+  element.click();
+  document.body.removeChild(element);
+}
+
+export function downloadMembers() {
+  api.Memberships
+    .all({ active: new Date() }, true)
+    .then(body => {
+      return body.reduce( (a, m) => {
+        if (a.indexOf(m.userDce) < 0) {
+          a.push(m.userDce);
+        }
+        return a;
+      }, []);
+    })
+    .then( members => {
+      saveAs(members.join('\n'), 'memberships.csv');
+    });
+}
+
 function getMembersFailure(error) {
   return {
     type: GET_MEMBERS_FAILURE,
