+++ app/js/containers/app.jsx
@@ -6,7 +6,7 @@ import { signOut, checkLogin } from '../actions/auth';
 import AddModal from '../components/add-modal';
 import { addMembership } from '../actions/memberships';
 import { getCommittees } from '../actions/committees';
-import API from '../api';
+import { downloadMembers } from '../actions/members';
 
 function mapStateToProps(state) {
   return {
@@ -28,8 +28,6 @@ class ScoreboardApp extends React.Component {
     this.showAdd = this.showAdd.bind(this);
     this.hideAdd = this.hideAdd.bind(this);
     this.addMembership = this.addMembership.bind(this);
-    this.saveAs = this.saveAs.bind(this);
-    this.downloadFile = this.downloadFile.bind(this);
   }
 
   componentDidMount() {
@@ -96,38 +94,11 @@ class ScoreboardApp extends React.Component {
     return <li><span/></li>;
   }
 
-  downloadFile() {
-    API
-      .Memberships
-      .all({ active: new Date() }, true)
-      .then(body => {
-        return body.reduce( (a, m) => {
-          if (a.indexOf(m.userDce) < 0) {
-            a.push(m.userDce);
-          }
-          return a;
-        }, []);
-      })
-      .then( members => {
-        this.saveAs(members.join('\n'), 'memberships.csv');
-      });
-  }
-
-  saveAs(contents, filename) {
-    const element = document.createElement('a');
-    element.setAttribute('href', 'data:text/plain;charset=utf-8,' + encodeURIComponent(contents));
-    element.setAttribute('download', filename);
-    element.style.display = 'none';
-    document.body.appendChild(element);
-    element.click();
-    document.body.removeChild(element);
-  }
-
   renderDownload() {
     if (this.props.auth.signedIn) {
       return (
         <li>
-          <button className='btn btn-link' onClick={this.downloadFile}>Download list of members</button>
+          <button className='btn btn-link' onClick={downloadMembers}>Download list of members</button>
         </li>
       );
     }
